home aliases (KISS)

+++ init.sh
@@ -2,13 +2,21 @@
 
 CURRENT=`pwd`
 SOURCE=.dotfiles
-FILES=".aliases .bash_functions .bash_login .bash_logout .bash_profile .bashrc .curlrc .gemrc .gitconfig .irbrc .profile .pryrc .railsrc .rvmrc .screenrc .zlogin .zprofile .zshenv .zshrc"
+FILES=".aliases .bash_functions .bash_login .bash_logout .bash_profile .bashrc .curlrc .gemrc .gitconfig .home_aliases .irbrc .profile .pryrc .railsrc .rvmrc .screenrc .zlogin .zprofile .zshenv .zshrc"
+
+if [ ! -f $SOURCE/.home_aliases ]; then
+  cp ~/$SOURCE/.home_aliases.template ~/$SOURCE/.home_aliases
+fi
+
 if [ ! -f $SOURCE/.gitconfig ]; then
   cp ~/$SOURCE/.gitconfig.template ~/$SOURCE/.gitconfig
 fi
+
 for f in $FILES
 do
-  ln -sF $SOURCE/$f ~/
+  if [ ! -f ~/$f ]; then
+    ln -sF $SOURCE/$f ~/
+  fi
 done
 
 # Sublime Text 2 config file
